remove methods for getting raw trace data back from level and directshape since these were really the same method , add a generic method into element binder for this repeated task that returns the element and tuple
